pcli(tx): move auction module into `tx` module
