+++ crates/bin/pcli/src/command/tx/auction.rs
